Fixed Custom Content Filter C# example in order to work without modify the project
